make start_service_checker work under `set -e`
ida-hub-acceptance-tests#92 introduced the use of `set -e` for
aborting when commands unexpectedly fail.

However, start_service_checker had a curl command which was expected
to fail, at least until the service comes up.  Therefore, `set -e`
caused start_service_checker to bomb out if curl ever failed.

This commit changes the function to run curl inside an `if` block,
which causes bash not to abort under `set -e` if curl fails.

I happen to think that the if block is more readable than the
`&& break || sleep 5` thing that it replaced, too. :)