Rebuild of the build-local script
build-local.rb has been rebuilt to pull the needed repos if missing
using git.  It has also been setup to allow users to specify the number
of threads to use.  This should help GDS mac users who may encounter
issues with Docker on the mac which is backed by a virtualbox virtual
machine with 2GB of RAM by default.