Updates to the build and startup script mostly
So I've updated the docker-compose to use a named network which we can
going forward other optional componets to.  The plan is to be able to
add matomo as an optional component going forward.

Update the build-local.rb script with minor improvements.

Added to a dozzle option to the startup.sh to bring up Dozzle optionally
if a user isn't running Docker Desktop.