Introduce retries to the build script
Sometimes the integration tests will fail when building SAML-Soap-proxy
and SAML-proxy.  This seems to be due to a bottle next in available
system resources when this first tries to build.  Attempting to run
startup again most times seems to result in a successful build.  So
these changes attempt to retry the build if it fails so the user doesn't
have to.