BAU Better data directory handling on Linux
This update does two things.  The first is that it gives us better
control over the data directory.  It allows us to manually remove the
data directory and automatically removes it after two weeks to stop the
data directory from getting stale.

The other change is that we supply the Docker container with the userid
of the user who is currently running the container and updates ownership
of the data directory and env files as part of the build process.