Updated the help and added a cleanup option
This should make the script easier to use.  I'll raise a seperate PR to
update the README.md later as this is a bigger task.  Also added a clean
task which cleans up the data directory and .env files.  This is much
more useful on Linux where these files are created as root.