Removed non-working scripts, broke apart ruby file
This is a bit of a big change.  I have removed the following scripts as
they don't seem to working any more:

  * restart.rb
  * shutdown-jars.sh

Updated the shutdown.sh script to actually shutdown dozzle if its
running and added an option to override this behaviour.

Created a new shell script for component management to wrap originally
around build-local.rb but now wraps the new component.rb script (see
below).

Broke apart the build-local.rb.  First I moved the class files in to
their own files.  Next I seperated out the new code for managing
components in to a components.rb script.  Finally I did a bit of a clean
up and moved all the ruby files in to the lib dir.  You shouldn't need
to access these directly as we have wrapper shell scripts for this.