BAU: Allow using SNAPSHOTS in maven local
Often when developing the hub we want to use a snapshot of another
dependency, which has been published to your local maven repo.

This change makes your local maven repo available in the docker builds.
It copies the repo with a namespace to avoid issues with multiple
threads.

It adds `COPY` lines to the Dockerfile on the fly, rather than having to
maintain a separate Dockerfile.