BAU Added experimental powershell scripts
These largely replicate their bashshell equivelenets but with a
focus on making things run natively in Windows (No WSL Required).
The script checks to make sure Docker, Git and Ruby are all installed
and if they are missing tells the user to go download them.

This is an inital commit of the PowerShell scripts.  Although they
have been extensively tested on Windows they've not yet been tested
using PowerShell running on MacOS or Linux.