BAU Updates to the repos.yml for ARM Hardware
The updates to the repos.yml file overrides the images used by the
verify hub so they are not pinned to a specific hardware arch which
should allow verify-local-startup to run on on M1 MacBooks and the 8GB
Raspberry Pi 4.  Also moved `verify-frontend` to the top of the
repos.yml file as it takes the longest to build on the Pi.  Current Pi
build time is around 7 - 8 hours from scratch so best this is done
overnight.